Move 'Confirm Modal' file location
